+++ gradle-plugins/react/brownfield/src/main/kotlin/com/callstack/react/brownfield/utils/DirectoryManager.kt
@@ -2,7 +2,6 @@ package com.callstack.react.brownfield.utils
 
 import com.callstack.react.brownfield.shared.BaseProject
 import com.callstack.react.brownfield.shared.Constants.INTERMEDIATES_TEMP_DIR
-import com.callstack.react.brownfield.shared.Constants.RE_BUNDLE_FOLDER
 import java.io.File
 
 object DirectoryManager : BaseProject() {
@@ -12,9 +11,5 @@ object DirectoryManager : BaseProject() {
 
     fun getKotlinMetaDirectory(variantName: String): File {
         return project.file("$buildDir/tmp/kotlin-classes/$variantName/META-INF")
-    }
-
-    fun getReBundleDirectory(variantName: String): File {
-        return project.file("$buildDir/outputs/$RE_BUNDLE_FOLDER/$variantName")
     }
 }